requirements.txt
colorama==0.4.6
matplotlib==3.10.0
numpy==2.2.1
pandas==2.2.3
pillow==10.4.0
plotly==5.22.0
pymia==0.3.2
pyyaml==6.0.2
scikit-image==0.25.0
scipy==1.15.0
simpleitk==2.3.1
statsmodels==0.14.0
stqdm==0.0.5
streamlit==1.41.1
streamlit-plotly-events==0.0.6
tqdm==4.64.0
loguru==0.7.2
click==8.0.0
kaleido==0.2.1
